Added access to the OCDB and simplify CVMFS
The OCDB is now available over CVMFS. The documentation has been
updated.

The cvmfs_chksetup is no longer called each time CVMFS is mounted. It
does not seem to be necessary, but it slows down start up significantly.
It leaves open the possibility of network issues being more difficult to
identify, but this has not yet been an issue, so it seems to be a
worthwhile change on balance.